chore: dsl-filter npm package
