+++ .babelrc
@@ -11,7 +11,6 @@
           // outside of this repo.
           "@extension-kid/core": "./node_modules/@extension-kid/core",
           "@extension-kid": "./packages/@extension-kid",
-          "@d2iq": "./packages/@d2iq",
           "PluginSDK": "./src/js/plugin-bridge/PluginSDK",
           "PluginTestUtils": "./src/js/plugin-bridge/PluginTestUtils",
           "#PLUGINS": "./plugins",
