--- packages/@d2iq/dsl-filter/index.js
@@ -1,32 +1,0 @@
-import * as SearchDSL from "./src/grammar/SearchDSL";
-import * as DSLCombinerTypes from "./src/DSLCombinerTypes";
-import * as DSLUtil from "./src/DSLUtil";
-import * as DSLUpdateUtil from "./src/DSLUpdateUtil";
-import * as DSLFormUtil from "./src/DSLFormUtil";
-import * as DSLParserUtil from "./src/DSLParserUtil";
-
-import { default as DSLFilterTypes, FilterType } from "./src/DSLFilterTypes";
-import { default as DSLExpression } from "./src/DSLExpression";
-import { default as DSLExpressionPart } from "./src/DSLExpressionPart";
-import { default as DSLUpdatePolicy } from "./src/DSLUpdatePolicy";
-import { ASTNode, CombinerNode, FilterNode } from "./src/DSLASTNodes";
-
-import DSLFilter from "./src/DSLFilter";
-
-const DSLASTNodes = { ASTNode, CombinerNode, FilterNode };
-
-export {
-  SearchDSL,
-  DSLASTNodes,
-  DSLFilter,
-  DSLFilterTypes,
-  FilterType,
-  DSLCombinerTypes,
-  DSLExpression,
-  DSLExpressionPart,
-  DSLFormUtil,
-  DSLParserUtil,
-  DSLUpdatePolicy,
-  DSLUpdateUtil,
-  DSLUtil
-};
