--- packages/@d2iq/dsl-filter/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@d2iq/dsl-filter",
-  "main": "index.js",
-  "version": "0.1.0",
-  "license": "Apache-2.0",
-  "scripts": {
-    "test": "jest",
-    "test:watch": "npm test -- --watch"
-  },
-  "peerDependencies": {},
-  "devDependencies": {},
-  "jest-webpack-alias": {},
-  "dependencies": {
-    "@types/jest": "22.2.3",
-    "jest-cli": "22.4.0",
-    "jison": "0.4.18",
-    "ts-jest": "22.4.6"
-  }
-}
