--- packages/@d2iq/dsl-filter/src/DSLCombinerTypes.ts
@@ -1,9 +1,0 @@
-/**
- * Different types of combiners (operators) in a DSL expression
- */
-const COMBINER_TYPES = {
-  AND: 1,
-  OR: 2
-};
-
-export default COMBINER_TYPES;
