--- packages/@d2iq/dsl-filter/src/DSLUpdatePolicy.js
@@ -1,9 +1,0 @@
-/**
- * Different policies for updating DSL attributes
- */
-const DSL_UPDATE_POLICY = {
-  Checkbox: Symbol("Checkbox"),
-  Radio: Symbol("Radio")
-};
-
-export default DSL_UPDATE_POLICY;
