+++ tsconfig.json
@@ -28,12 +28,11 @@
     "resolveJsonModule": true,
     "skipLibCheck": true,
     "paths": {
+      "@d2iq/dsl-filter": ["src/typings/d2iq__dsl-filter"],
       "@dcos/http-service": ["src/typings/dcos__http-service"],
       "@dcos/mesos-client": ["src/typings/dcos__mesos-client"],
       "@extension-kid/core": ["node_modules/@extension-kid/core"],
       "@extension-kid/*": ["packages/@extension-kid/*"],
-      "@d2iq/*": ["packages/@d2iq/*"],
-      "@d2iq/dsl-filter": ["src/typings/d2iq__dsl-filter"],
       "*": ["*", "packages/*", "src/typings/*"],
       "#SRC/*": ["src/*"],
       "#LOCALE/*": ["locale/*"],
